+++ aimsalgo/src/AimsApplyTransform/AimsApplyTransform.cc
@@ -33,6 +33,8 @@
 #include <boost/algorithm/string/trim.hpp>
 #include <boost/lexical_cast.hpp>
 //----------------------------------------------------------------------------
+// #include <soma-io/writer/pythonwriter.h>
+
 
 using namespace aims;
 using namespace std;
@@ -90,6 +92,8 @@ public:
   string  progress_file;
   bool    volume_id;
   rc_ptr<TransformationGraph3d> trans_graph;
+  bool auto_reference;
+  Object reference_header;
 };
 
 
@@ -108,7 +112,8 @@ ApplyTransformProc::ApplyTransformProc()
     sx(0.), sy(0.), sz(0.),
     vfinterp("linear"),
     mmap_fields(false),
-    volume_id( false )
+    volume_id( false ),
+    auto_reference( false )
 {
   registerProcessType("Volume", "S8",      &doVolume<int8_t, int8_t>);
   registerProcessType("Volume", "U8",      &doVolume<uint8_t, uint8_t>);
@@ -213,11 +218,16 @@ void set_geometry_from_header(ApplyTransformProc& proc,
 }
 
 
-// Return the header of the reference object, or a null reference if there is
-// no reference object.
-carto::Object read_reference_header(const std::string reference_filename)
+// Set and return the header of the reference object, or a null reference if
+// there is no reference object.
+carto::Object read_reference_header( ApplyTransformProc & proc )
 {
-  carto::Object reference_header;
+  string reference_filename = proc.reference;
+  if( reference_filename == "auto" )
+  {
+    proc.auto_reference = true;
+    return proc.reference_header;
+  }
   if(!reference_filename.empty()) {
     Finder finder;
     if(!finder.check(reference_filename)) {
@@ -229,12 +239,12 @@ carto::Object read_reference_header(const std::string reference_filename)
       // graphs are their own header, thus we have to read them entirely
       Reader<Graph> r( reference_filename );
       Graph *g = r.read();
-      reference_header.reset( g );
+      proc.reference_header.reset( g );
     }
     else
-      reference_header = finder.headerObject();
+      proc.reference_header = finder.headerObject();
   }
-  return reference_header;
+  return proc.reference_header;
 }
 
 
@@ -242,7 +252,6 @@ carto::Object read_reference_header(const std::string reference_filename)
 // of priority, use commandline flags, else the reference object, else the
 // input object passed as fallback_header.
 void set_geometry(ApplyTransformProc& proc,
-                  const carto::Object& reference_header,
                   const DictionaryInterface& fallback_header,
                   bool set_dimensions)
 {
@@ -252,9 +261,10 @@ void set_geometry(ApplyTransformProc& proc,
     return;
 
   // Reference volume
-  if(!reference_header.isNull()) {
+  if(!proc.reference_header.isNull()) {
+    // cout << "REF HDR:\n";
     try {
-      set_geometry_from_header(proc, *reference_header, dimension_ok,
+      set_geometry_from_header(proc, *proc.reference_header, dimension_ok,
                                voxel_size_ok);
     } catch(...) {
       throw FatalError("Failed to retrieve volume_dimension and voxel_size "
@@ -391,7 +401,6 @@ insert_transformation_to_old_referential(DictionaryInterface& header,
 void adjust_header_transforms(const ApplyTransformProc& proc,
                               DictionaryInterface& header,
                               const Transformation3d * inverse_transform,
-                              const carto::Object reference_header,
                               const string & out_ref)
 {
   if(proc.keep_transforms) {
@@ -413,7 +422,7 @@ void adjust_header_transforms(const ApplyTransformProc& proc,
   const AffineTransformation3d * affine_inverse_transform
     = dynamic_cast<const AffineTransformation3d*>(inverse_transform);
 
-  if(!proc.ignore_reference_transforms && !reference_header.isNull()) {
+  if(!proc.ignore_reference_transforms && !proc.reference_header.isNull()) {
     // Case 2: copy the referential and transformations from the reference
     // object if enabled and available.
 
@@ -428,7 +437,7 @@ void adjust_header_transforms(const ApplyTransformProc& proc,
     std::string reference_referential;
     try {
       reference_referential =
-        reference_header->getProperty("referential")->getString();
+        proc.reference_header->getProperty("referential")->getString();
     } catch(...) {
     }
     string out_ref2 = out_ref;
@@ -443,8 +452,8 @@ void adjust_header_transforms(const ApplyTransformProc& proc,
     carto::Object new_referentials_obj, new_transforms_obj;
     try {
       copy_referentials_and_transformations(
-        reference_header->getProperty("referentials"),
-        reference_header->getProperty("transformations"),
+        proc.reference_header->getProperty("referentials"),
+        proc.reference_header->getProperty("transformations"),
         new_referentials_obj,
         new_transforms_obj,
         reference_referential, out_ref2, *proc.trans_graph
@@ -669,27 +678,25 @@ std::pair<std::pair<const_ref<Transformation3d>,
                     const_ref<Transformation3d> >, string>
 load_transformations(ApplyTransformProc& proc,
                      const DictionaryInterface* input_header = nullptr,
-                     const Finder & finder = Finder(),
-                     const Object reference_header = Object())
+                     const Finder & finder = Finder() )
 {
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> ret_with_ref;
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> > & ret
     = ret_with_ref.first;
 
-  rc_ptr<Transformation3d> aims_to_input_space_transform; // null
-
   // boost::iequals is used for case-insensitive comparison
   using boost::iequals;
   string input_coords = boost::trim_copy(proc.input_coords);
   string output_coords = boost::trim_copy(proc.output_coords);
   string output_space = boost::trim_copy(proc.output_space);
+  cout << "input_coords: " << input_coords << endl;
+  cout << "output coords: " << output_coords << endl;
+  cout << "output_space: " << output_space << endl;
 
   if(proc.points_mode) {
     if(iequals(input_coords, "auto"))
       input_coords = "AIMS";
-    else if(!iequals(input_coords, "AIMS"))
-      throw FatalError("--input-coords cannot be used in points mode");
   }
 
   rc_ptr<TransformationGraph3d> tg( 0 );
@@ -743,12 +750,16 @@ load_transformations(ApplyTransformProc& proc,
   }
   else
     trefs = tg->updateFromObjectHeader( input_header, true );
-  otrefs = tg->updateFromObjectHeader( reference_header, true );
+  otrefs = tg->updateFromObjectHeader( proc.reference_header, true );
   ref = trefs[0];
 
   // determine input and output spaces
 
-  Vertex *rv1 = 0, *rv2 = 0, *rv3 = 0;  // referentials in graph
+  Vertex *rv0 = 0, *rv1 = 0, *rv2 = 0, *rv3 = 0;  // referentials in graph
+  // rv0: input (aims object) space
+  // rv1: space of source of direct transform
+  // rv2: space of destg of direct transform
+  // rv3: output (aims object) space
   try
   {
     rv1 = tg->referentialByCode( input_coords, input_header, trefs );
@@ -761,9 +772,29 @@ load_transformations(ApplyTransformProc& proc,
   if( !rv1 )
     throw FatalError( "Could not find the input space referential" );
 
+  if( input_header && input_header->hasProperty( "referential" ) )
+  {
+    try
+    {
+      rv0 = tg->referentialByCode(
+        input_header->getProperty( "referential" )->getString(), input_header,
+                                  trefs );
+    }
+    catch( runtime_error & )
+    {
+    }
+  }
+  if( !rv0 )
+    rv0 = rv1;
+  string iref = rv0->getProperty( "uuid" )->getString();
+
+  ref = rv1->getProperty( "uuid" )->getString();
+  // cout << "input ref: " << ref << endl;
+
   try
   {
-    rv2 = tg->referentialByCode( output_coords, reference_header, otrefs );
+    rv2 = tg->referentialByCode( output_coords, proc.reference_header,
+                                 otrefs );
   }
   catch( runtime_error & e )
   {
@@ -775,9 +806,70 @@ load_transformations(ApplyTransformProc& proc,
     throw FatalError(
       "Could not find the output space (--output-coords) referential" );
 
+  if( output_space == "AIMS" && proc.auto_reference )
+  {
+    // look for a header in the transform chain
+    // cout << "look for output_space\n";
+    Vertex::const_iterator ie, ee = rv2->end();
+    for( ie=rv2->begin(); ie!=ee; ++ie )
+      if( *(*ie)->begin() == rv2 )
+      {
+        rc_ptr<Transformation3d> t = tg->transformation( *ie );
+
+        if( !t.get() && (*ie)->hasProperty( "filename" ) )
+        {
+          string filename = (*ie)->getProperty( "filename" )->getString();
+          if( ( filename.length() > 4
+                && filename.substr( filename.length() - 4, 4 ) == ".trm" )
+              || ( filename.length() > 10
+                   && filename.substr( filename.length() - 10, 10 )
+                     == ".trm?inv=1" ) )
+          {
+            Edge *e = tg->getTransformation(
+              rv2->getProperty("uuid")->getString(),
+              (*(*ie)->rbegin())->getProperty("uuid")->getString() );
+            t = tg->transformation( e );
+          }
+        }
+
+        if( t.get() && dynamic_cast<AffineTransformation3d *>( t.get() ) )
+        {
+          AffineTransformation3d *at = dynamic_cast<AffineTransformation3d *>(
+            t.get() );
+          bool match = true;
+          for( int y=0; y<3; ++y )
+            for( int x=0; x<3; ++x )
+            {
+              double z = std::abs( at->affine().at( x, y ) );
+              if( x == y )
+              {
+                if( z < 0.9999 || z > 1.0001 )
+                {
+                  match = false;
+                  break;
+                }
+              }
+              else if( z > 0.0001 )
+              {
+                match = false;
+                break;
+              }
+            }
+          if( match )
+          {
+            output_space = (*(*ie)->rbegin())
+              ->getProperty( "uuid" )->getString();
+            cout << "found AIMS output_space: " << output_space << endl;
+            proc.output_space = output_space;
+            break;
+          }
+        }
+      }
+  }
+
   try
   {
-    rv3 = tg->referentialByCode( output_space, reference_header, otrefs );
+    rv3 = tg->referentialByCode( output_space, proc.reference_header, otrefs );
   }
   catch( runtime_error & e )
   {
@@ -789,8 +881,10 @@ load_transformations(ApplyTransformProc& proc,
     throw FatalError(
       "Could not find the output space (--output-space) referential" );
   oref = rv3->getProperty( "uuid" )->getString();
+  proc.output_space = oref;
   ret_with_ref.second = oref;
   bool has_dir = false;
+  // cout << "oref: " << oref << endl;
 
   if(!proc.direct_transform_list.empty())
   {
@@ -805,6 +899,7 @@ load_transformations(ApplyTransformProc& proc,
         = load_transformation(*filename_it, proc);
       direct_chain.push_back(transform);
     }
+
     const_ref<Transformation3d> tcr = direct_chain.simplify();
     // indirect convert const_ref -> rc_ptr (non-const)
     rc_ptr<Transformation3d> tc( const_cast<Transformation3d *>(
@@ -851,20 +946,20 @@ load_transformations(ApplyTransformProc& proc,
   tg->registerInverseTransformations();
 
   // now get the complete transform chains
-  Edge *tde = tg->getTransformation( ref, oref, true );
-  Edge *tie = tg->getTransformation( oref, ref, true );
+  Edge *tde = tg->getTransformation( iref, oref, true );
+  Edge *tie = tg->getTransformation( oref, iref, true );
   if( tde )
     ret.first = tg->transformation( tde );
   if( tie )
     ret.second = tg->transformation( tie );
 
-  // cout << "use direct " << ref << "->" << oref << ": " << ret.first.pointer() << ": " << typeid(*ret.first).name() << endl;
-  // cout << "use inv " << oref << "->" << ref << ": " << ret.second.pointer() << ": " << typeid(*ret.second).name() << endl;
+  // cout << "use direct " << iref << "->" << oref << ": " << ret.first.pointer() << ": " << typeid(*ret.first).name() << endl;
+  // cout << "use inv " << oref << "->" << iref << ": " << ret.second.pointer() << ": " << typeid(*ret.second).name() << endl;
 
   if(ret.first.isNull() && ret.second.isNull())
   {
     // get aims -> input space, if it exists, assume this is the one
-    tde = tg->getTransformation( tg->referentialById( ref ), rv1, true );
+    tde = tg->getTransformation( tg->referentialById( iref ), rv1, true );
     // cout << "use aims -> input: " << tde << endl;
     if( tde )
       ret.first = tg->transformation( tde );
@@ -909,7 +1004,7 @@ load_transformations(ApplyTransformProc& proc,
         vs[0] = vs[0] - proc.sx;
         vs[1] = vs[1] - proc.sy;
         vs[2] = vs[2] - proc.sz;
-        cout << "translation: " << vs << endl;
+        // cout << "translation: " << vs << endl;
 
         AffineTransformation3d *t = new AffineTransformation3d;
         rc_ptr<Transformation3d> tid( t );
@@ -939,6 +1034,89 @@ load_transformations(ApplyTransformProc& proc,
     ret.second = const_ref<Transformation3d>( ti3 );
   }
 
+  if( proc.auto_reference )
+  {
+    // get reference from last field
+    if( !ret.second.isNull() )
+    {
+      // cout << "auto ref: look for reference from inverse trans\n";
+      output_coords = rv2->getProperty( "uuid" )->getString();
+      Edge *tis = tg->getTransformation( output_coords, iref, true );
+      if( tis )
+      {
+        const_ref<Transformation3d> ts = tg->transformation( tis );
+        if( ts->header()->hasProperty( "voxel_size" ) )
+        {
+          cout << "auto ref: set reference from inverse trans\n";
+          if( proc.reference_header.isNull() )
+            proc.reference_header = Object::value( Dictionary() );
+          proc.reference_header->copyProperties( ts->header() );
+        }
+        else if( dynamic_cast<const TransformationChain3d *>( ts.pointer() ) )
+        {
+          const TransformationChain3d *tc
+            = dynamic_cast<const TransformationChain3d *>( ts.pointer() );
+          Object auto_hdr;
+          for( auto it=tc->begin(); it!=tc->end(); ++it )
+            if( !(*it)->header().isNull()
+                && (*it)->header()->hasProperty( "voxel_size" ) )
+            {
+              auto_hdr = (*it)->header();
+              break;
+            }
+
+          if( !auto_hdr.isNull() )
+          {
+            // cout << "auto ref: set reference from transform chain\n";
+            if( proc.reference_header.isNull() )
+              proc.reference_header = Object::value( Dictionary() );
+            proc.reference_header->copyProperties( auto_hdr );
+            if( proc.reference_header->hasProperty( "storage_to_memory" ) )
+              proc.reference_header->removeProperty( "storage_to_memory" );
+
+            // this referential is the "output-coords" space
+            proc.reference_header->setProperty( "referential", oref );
+
+            if( oref != output_coords )
+            {
+
+              tis = tg->getTransformation( oref, output_coords, true );
+              if( tis )
+              {
+                ts = tg->transformation( tis );
+                if( dynamic_cast<const TransformationChain3d *>(
+                    ts.pointer() ) )
+                  ts = dynamic_cast<const TransformationChain3d *>(
+                    ts.pointer() )->simplify();
+                if( dynamic_cast<const AffineTransformation3d *>(
+                    ts.pointer() ) )
+                {
+                  const AffineTransformation3d *at
+                    = dynamic_cast<const AffineTransformation3d *>(
+                      ts.pointer() );
+                  // cout << "set in reference new transform to " << proc.output_coords << endl;
+                  vector<vector<float> > reftrans( 1 );
+                  reftrans[0] = at->toVector();
+                  proc.reference_header->setProperty( "transformations",
+                                                      reftrans );
+                  vector<string> refrefs( 1 );
+                  refrefs[0] = proc.output_coords;
+                  proc.reference_header->setProperty( "referentials",
+                                                      refrefs );
+                  // cout << "done.\n";
+                  // PythonWriter pw;
+                  // pw.attach(cout);
+                  // pw.write(proc.reference_header);
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+
+  }
+
   return ret_with_ref;
 }
 
@@ -959,15 +1137,14 @@ bool doVolume(Process & process, const string & fileref, Finder & finder)
   stringTo(proc.background_value, background_value);
 
   // Compute dimensions of the output volume
-  const carto::Object reference_header = read_reference_header(proc.reference);
-  set_geometry(proc, reference_header, input_image.header(), true);
+  read_reference_header(proc);
 
   // Load the transformation
   const_ref<Transformation3d> inverse_transform;
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> transforms_ref
-    = load_transformations(proc, &input_image.header(), finder,
-                            reference_header);
+    = load_transformations(proc, &input_image.header(), finder);
+  set_geometry(proc, input_image.header(), true);
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> >
     transforms = transforms_ref.first;
   inverse_transform = transforms.second;
@@ -1005,7 +1182,7 @@ bool doVolume(Process & process, const string & fileref, Finder & finder)
   }
 
   adjust_header_transforms(proc, out.header(), inverse_transform.pointer(),
-                           reference_header, transforms_ref.second);
+                           transforms_ref.second);
   // The UUID should NOT be preserved, because the output is a different file
   // from the input. Keeping it triggers the infamous "duplicate UUID" problem
   // in BrainVISA/Axon...
@@ -1018,6 +1195,10 @@ bool doVolume(Process & process, const string & fileref, Finder & finder)
   resampler->resample_inv(input_image, *inverse_transform,
                           background_value, out, true);
   cout << endl;
+
+  // PythonWriter pw;
+  // pw.attach( cout );
+  // pw.write(Object::reference(out.header()));
 
   // Write the resampled volume. The allow_orientation_change instructs the
   // NIfTI writer to avoid writing dummy transformations to the header of the
@@ -1045,13 +1226,13 @@ bool doMesh(Process & process, const string & fileref, Finder & finder)
   AimsTimeSurface<D, Void> mesh;
   input_reader.read(mesh);
 
-  const carto::Object reference_header = read_reference_header(proc.reference);
+  read_reference_header(proc);
 
   // Load the transformation
   const_ref<Transformation3d> direct_transform, inverse_transform;
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> transforms_ref
-    = load_transformations(proc, &mesh.header(), finder, reference_header);
+    = load_transformations(proc, &mesh.header(), finder);
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> >
     transforms = transforms_ref.first;
   direct_transform = transforms.first;
@@ -1062,7 +1243,7 @@ bool doMesh(Process & process, const string & fileref, Finder & finder)
   }
 
   adjust_header_transforms(proc, mesh.header(), inverse_transform.pointer(),
-                           reference_header, transforms_ref.second);
+                           transforms_ref.second);
   // The UUID should NOT be preserved, because the output is a different file
   // from the input. Keeping it triggers the infamous "duplicate UUID" problem
   // in BrainVISA/Axon...
@@ -1097,8 +1278,8 @@ bool doBucket(Process & process, const string & fileref, Finder & finder)
   input_reader.read(input_bucket);
 
   // Prepare the output dimensions
-  const carto::Object reference_header = read_reference_header(proc.reference);
-  set_geometry(proc, reference_header, input_bucket.header(), false);
+  read_reference_header(proc);
+  set_geometry(proc, input_bucket.header(), false);
   cout << "Output voxel size: "
        << proc.sx << ", " << proc.sy << ", " << proc.sz << " mm" << endl;
 
@@ -1106,8 +1287,7 @@ bool doBucket(Process & process, const string & fileref, Finder & finder)
   const_ref<Transformation3d> direct_transform, inverse_transform;
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> transforms_ref
-    = load_transformations(proc, &input_bucket.header(), finder,
-                            reference_header);
+    = load_transformations(proc, &input_bucket.header(), finder);
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> >
     transforms = transforms_ref.first;
   direct_transform = transforms.first;
@@ -1145,7 +1325,7 @@ bool doBucket(Process & process, const string & fileref, Finder & finder)
   out->header().setProperty( "referential", transforms_ref.second );
 
   adjust_header_transforms(proc, out->header(), inverse_transform.pointer(),
-                           reference_header, transforms_ref.second);
+                           transforms_ref.second);
 
   // Write the resampled volume
   Writer<BucketMap<Void> > w3(proc.output);
@@ -1162,14 +1342,13 @@ bool doBundles(Process & process, const string & fileref, Finder & finder)
   // Prepare the Bundle reader
   aims::BundleReader bundle_reader(fileref);
 
-  const carto::Object reference_header = read_reference_header(proc.reference);
+  read_reference_header(proc);
 
   // Load the transformation
   const_ref<Transformation3d> direct_transform, inverse_transform;
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> transforms_ref
-    = load_transformations(proc, bundle_reader.readHeader().get(), finder,
-                            reference_header);
+    = load_transformations(proc, bundle_reader.readHeader().get(), finder);
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> >
     transforms = transforms_ref.first;
   direct_transform = transforms.first;
@@ -1223,8 +1402,8 @@ bool doGraph(Process & process, const string & fileref, Finder & f)
   unique_ptr<Graph> graph(input_reader.read());
 
   // Deduce the voxel size of the output Graph
-  const carto::Object reference_header = read_reference_header(proc.reference);
-  set_geometry(proc, reference_header, *graph, false);
+  read_reference_header(proc);
+  set_geometry(proc, *graph, false);
   cout << "Output voxel size: "
        << proc.sx << ", " << proc.sy << ", " << proc.sz << " mm" << endl;
 
@@ -1233,7 +1412,7 @@ bool doGraph(Process & process, const string & fileref, Finder & f)
   const_ref<Transformation3d> direct_transform, inverse_transform;
   std::pair<std::pair<const_ref<Transformation3d>,
                       const_ref<Transformation3d> >, string> transforms_ref
-    = load_transformations(proc, graph.get(), f, reference_header);
+    = load_transformations(proc, graph.get(), f);
   std::pair<const_ref<Transformation3d>, const_ref<Transformation3d> >
     transforms = transforms_ref.first;
   direct_transform = transforms.first;
@@ -1245,7 +1424,7 @@ bool doGraph(Process & process, const string & fileref, Finder & f)
 
   adjust_header_transforms(proc, *graph,
                            inverse_transform.pointer(),
-                           reference_header, transforms_ref.second);
+                           transforms_ref.second);
   // Update the transformation to Talairach stored in the old attributes
   // (Talairach_rotation, Talairach_translation, and Talairach_scale) to
   // reflect the updated transformations.
@@ -1511,14 +1690,12 @@ int main(int argc, const char **argv)
     app.addOption(proc.output_coords, "--output-coords",
                   "Output space identifier, or how to interpret coordinates "
                   "in the output image w.r.t. the transformations written in "
-                  "the reference image header. Note that this parameter "
+                  "the reference image header (thus, the destination of this "
+                  "transformation). Note that this parameter "
                   "specifies the destination space of the transformations "
                   "passed (using -d / -I), but *not* the space we want to "
                   "produce outputs for: they may differ - see the "
-                  "--output-space option for this. Especially, if -d/-I "
-                  "options are not used (if a transformations graph is "
-                  "specified or if transformations information is in data "
-                  "headers), then this option is not used at all. See above. "
+                  "--output-space option for this. "
                   "[default: AIMS]",
                   true);
     app.addOption(proc.output_space, "--output-space",
@@ -1562,7 +1739,11 @@ int main(int argc, const char **argv)
     app.addOption(proc.reference, "--reference",
                   "Volume used to define output voxel size and volume "
                   "dimension (values are overridden by --dx, --dy, "
-                  "--dz, --sx, --sy and --sz)", true);
+                  "--dz, --sx, --sy and --sz). If \"auto\" and if the last "
+                  "transformation in the chain is a "
+                  "deformation field which has a field of view, then use this "
+                  "one. It is normally used when --output-coords is "
+                  "specified, and --output-space is left to AIMS.", true);
     app.addOption(proc.keep_transforms, "--keep-transforms",
                   "Preserve the transformations of the input image", true);
     app.addOption(proc.ignore_reference_transforms,